Use the `wait_timer` to target the `warning_delay`.
Since `wait_timer` is likely unique among `autolight` automations, this
will allow HASS to trigger only the minimal set of automation and
prevent log polution. We still check that the response is being sent to
the current automation. Ideally, we would target the automation ID
directly, but it's not possible to do that, since we can't use templates
in trigger definitions.